[#7971] export to_datastore_plugin_data validator
