+++ ckanext/example_idatadictionaryform/plugin.py
@@ -2,10 +2,8 @@
 
 from __future__ import annotations
 
-from typing import Any
-from ckan.types import (
-    Schema, Context, FlattenErrorDict, FlattenDataDict, FlattenKey
-)
+from typing import Any, cast
+from ckan.types import Schema, ValidatorFactory
 from ckan.common import CKANConfig
 
 import json
@@ -29,16 +27,20 @@ class ExampleIDataDictionaryFormPlugin(plugins.SingletonPlugin):
     def update_datastore_create_schema(self, schema: Schema):
         ignore_empty = get_validator('ignore_empty')
         int_validator = get_validator('int_validator')
+        to_datastore_plugin_data = cast(
+            ValidatorFactory, get_validator('to_datastore_plugin_data'))
+        to_eg_iddf = to_datastore_plugin_data('example_idatadictionaryform')
 
-        assert isinstance(schema['fields'], dict)
-        f = schema['fields']
-        f['an_int'] = [ignore_empty, int_validator, to_plugin_data()]
-        f['json_obj'] = [ignore_empty, json_obj, to_plugin_data()]
+        f = cast(Schema, schema['fields'])
+        f['an_int'] = [ignore_empty, int_validator, to_eg_iddf]
+        f['json_obj'] = [ignore_empty, json_obj, to_eg_iddf]
+
         # use different plugin_key so that value isn't removed
         # when above fields are updated
         f['secret'] = [
             ignore_empty,
-            to_plugin_data('example_idatadictionaryform_secrets')]
+            to_datastore_plugin_data('example_idatadictionaryform_secret')
+        ]
         return schema
 
     def update_datastore_info_field(
@@ -59,21 +61,3 @@ def json_obj(value: str | dict[str, Any]) -> dict[str, Any]:
         return value
     except (TypeError, ValueError):
         raise Invalid('Not a JSON object')
-
-
-def to_plugin_data(plugin_key: str = 'example_idatadictionaryform'):
-    def validator(
-            key: FlattenKey,
-            data: FlattenDataDict,
-            errors: FlattenErrorDict,
-            context: Context):
-        """
-        move value to plugin_data dict
-        """
-        value = data.pop(key)
-        field_index = key[-2]
-        field_name = key[-1]
-        context['plugin_data'].setdefault(
-            field_index, {}).setdefault(
-            plugin_key, {})[field_name] = value
-    return validator
